refactor(3d): move planar-pose out of kornia-apriltag; publish kornia-calib
`estimate_tag_pose` lived in kornia-apriltag but its imports were only
kornia_3d::pose::{homography_4pt2d, Pose3d} and kornia_algebra. It is
Lu-Hager-Mjolsness orthogonal iteration over 4 coplanar correspondences --
nothing about it is AprilTag-specific. A tag is one caller among many.

Because it sat there, every consumer of planar pose had to depend on the tag
decoder. kornia-calib in particular carried a hard dependency on
kornia-apriltag for exactly ONE import, and since kornia-apriltag is excluded
from crates.io (large git submodule of test images) that made kornia-calib
unpublishable too -- a packaging accident, not a design constraint.

Moves the solver to `kornia_3d::pose::planar` as `estimate_planar_pose`, with
`PlanarPose`/`PlanarPosePair`/`PlanarPoseError`. kornia-apriltag keeps a thin
re-export under the old names so existing callers compile unchanged.

kornia-calib then drops kornia-apriltag entirely (verified: absent from
`cargo tree -e normal`, transitively too), so `publish = false` is removed and
it joins PUBLISH_ORDER.

The module doc is rewritten for the new home, including a caveat the old one
did not state: the initial homography is fitted against a canonical unit
square, so initialisation is exact only for square targets in that corner
order. The orthogonal iteration then refines against the caller's real
object points, so non-square coplanar quads still converge -- they just start
further away and may want more iterations.
