+++ crates/kornia-calib/Cargo.toml
@@ -9,9 +9,6 @@ homepage = { workspace = true }
 repository = { workspace = true }
 license = { workspace = true }
 categories = { workspace = true }
-# Not published: transitively depends on kornia-apriltag, which is excluded from
-# crates.io (large git submodule of test images). See scripts/release_rust.sh.
-publish = false
 
 [[bench]]
 harness = false
@@ -20,7 +17,6 @@ name = "bench_calibrate"
 [dependencies]
 kornia-3d = { workspace = true }
 kornia-algebra = { workspace = true }
-kornia-apriltag = { workspace = true }
 log = { workspace = true }
 thiserror = { workspace = true }
